+++ backend-v2/src/main/java/pt/ipleiria/careline/domain/dto/HealthData/TemperatureDTO.java
@@ -5,21 +5,21 @@ import pt.ipleiria.careline.domain.dto.PatientDTO;
 import java.time.Instant;
 
 public class TemperatureDTO extends HealthDataDTO {
-    private Integer temperature;
+    private Float temperature;
 
     public TemperatureDTO() {
     }
 
-    public TemperatureDTO(PatientDTO patient, Instant createdAt, Integer temperature) {
+    public TemperatureDTO(PatientDTO patient, Instant createdAt, Float temperature) {
         super(patient, createdAt);
         this.temperature = temperature;
     }
 
-    public Integer getTemperature() {
+    public Float getTemperature() {
         return temperature;
     }
 
-    public void setTemperature(Integer temperature) {
+    public void setTemperature(Float temperature) {
         this.temperature = temperature;
     }
